// server/api/src/user/profile-title.ts
export enum ProfileTitle {
  FIRSTNAME = 'firstname',
  LASTNAME = 'lastname',
  AVATAR_URL = 'avatarUrl',
  PHOTO = 'photo',
  LOCATION = 'location',
  TIME_ZONE = 'timeZone',
  COMPANY = 'company',
  GOOGLE = 'google',
  LINKEDIN = 'linkedin',
  MICROSOFT = 'microsoft',
  OTHER_EMAIL = 'other_email',
}