// server/api/src/user/profile-type.ts
export enum ProfileType {
  FIRSTNAME = 'firstname',
  LASTNAME = 'lastname',
  AVATAR_URL = 'avatarUrl',
  PHOTO = 'photo',
  LOCATION = 'location',
  TIME_ZONE = 'timeZone', // Timezone Identifier as defined in https://en.wikipedia.org/wiki/List_of_tz_database_time_zones - eg: "Africa/Bamako"
  COMPANY = 'company',
  EMAIL = 'email',
  TOKEN = 'token',
}